+++ tests/testthat/test_MsBackendSql.R
@@ -1,8 +1,9 @@
 test_that("backendInitialize works", {
     expect_error(backendInitialize(MsBackendSql()), "required")
-    expect_error(backendInitialize(MsBackendSql(), dbcon = "file"), "connection")
+    expect_error(backendInitialize(MsBackendSql(), dbcon = "file"),
+                 "connection")
 
-    be <- backendInitialize(MsBackendSql(), dbcon = mm8_db)
+    be <- backendInitialize(MsBackendSql(), dbcon = mm8_db_long)
 
     ## backendInitialize creating a new database.
     expect_warning(
@@ -27,37 +28,37 @@ test_that("dataStorage works", {
     res <- dataStorage(MsBackendSql())
     expect_identical(res, character())
 
-    res <- dataStorage(mm8_be)
+    res <- dataStorage(mm8_be_blob)
     expect_true(is.character(res))
-    expect_identical(length(res), length(mm8_be))
+    expect_identical(length(res), length(mm8_be_blob))
 })
 
 test_that("extractByIndex,MsBackendSql works", {
     idx <- c(4L, 12L, 100L, 14L)
-    res <- extractByIndex(mm8_be, idx)
+    res <- extractByIndex(mm8_be_blob, idx)
     expect_identical(res@spectraIds, idx)
 })
 
 test_that("[,MsBackendSql works", {
-    res <- mm8_be[]
-    expect_equal(res, mm8_be)
+    res <- mm8_be_blob[]
+    expect_equal(res, mm8_be_blob)
 
     idx <- c(4L, 12L, 100L, 14L)
-    res <- mm8_be[idx]
+    res <- mm8_be_blob[idx]
     expect_identical(res@spectraIds, idx)
 
     ## Duplicated elements.
     idx <- c(2L, 5L, 1L, 2L, 5L)
-    res <- mm8_be[idx]
+    res <- mm8_be_blob[idx]
     expect_identical(res@spectraIds, idx)
 
     ## With additional data.
-    be <- mm8_be
+    be <- mm8_be_blob
     be$new_var <- 1:length(be)
     res <- be[idx]
     expect_identical(be$new_var[idx], res$new_var)
 
-    res <- peaksData(mm8_be[c(3, 1)])
+    res <- peaksData(mm8_be_long[c(3, 1)])
     expect_true(length(res) == 2L)
     expect_equal(res[[1L]], peaksData(mm8_sps[3L])[[1L]])
     expect_equal(res[[2L]], peaksData(mm8_sps[1L])[[1L]])
@@ -67,7 +68,7 @@ test_that("[,MsBackendSql works", {
     expect_equal(res[[1L]], peaksData(mm8_sps[3L])[[1L]])
     expect_equal(res[[2L]], peaksData(mm8_sps[1L])[[1L]])
 
-    res <- peaksData(mm8_be[1L])
+    res <- peaksData(mm8_be_long[1L])
     expect_true(length(res) == 1L)
     expect_true(is.matrix(res[[1L]]))
     expect_equal(res[[1L]], peaksData(mm8_sps[1L])[[1L]])
@@ -77,30 +78,29 @@ test_that("[,MsBackendSql works", {
     expect_true(is.matrix(res[[1L]]))
     expect_equal(res[[1L]], peaksData(mm8_sps[1L])[[1L]])
 
-    tmp <- mm8_be
+    tmp <- mm8_be_long
     tmp@spectraIds <- c(tmp@spectraIds, 300L)
     res <- peaksData(tmp)
     res <- res[[length(res)]]
     expect_true(is.matrix(res))
     expect_true(nrow(res) == 0)
-
 })
 
 test_that("peaksData,MsBackendSql works", {
     idx <- c(4L, 12L, 100L, 14L)
-    res <- mm8_be[idx]
+    res <- mm8_be_long[idx]
     expect_identical(peaksData(res), peaksData(mm8_sps@backend[idx]))
 
     idx <- c(2L, 5L, 1L, 2L, 5L)
-    res <- mm8_be[idx]
+    res <- mm8_be_long[idx]
     expect_identical(peaksData(res), peaksData(mm8_sps@backend[idx]))
 
-    res <- peaksData(mm8_be)
-    res_2 <- peaksData(mm8_be, "mz")
+    res <- peaksData(mm8_be_long)
+    res_2 <- peaksData(mm8_be_long, "mz")
     expect_true(colnames(res_2[[1]]) == "mz")
     expect_equal(res[[1]][, "mz"], res_2[[1]][, "mz"])
 
-    res_2 <- peaksData(mm8_be, c("intensity", "mz"))
+    res_2 <- peaksData(mm8_be_long, c("intensity", "mz"))
     expect_true(all(colnames(res_2[[1]]) == c("intensity", "mz")))
     expect_equal(res[[1]][, "mz"], res_2[[1]][, "mz"])
     expect_equal(res[[1]][, "intensity"], res_2[[1]][, "intensity"])
@@ -127,23 +127,34 @@ test_that("peaksData,MsBackendSql works", {
     res_2 <- peaksData(mm8_be_blob[idx], c("intensity"))
     expect_equal(res_2, peaksData(mm8_sps@backend[idx], "intensity"))
 
-    res <- peaksData(mm8_be[1L])
+    res <- peaksData(mm8_be_long[1L])
     expect_true(is.list(res))
     expect_true(length(res) == 1L)
     expect_true(is.matrix(res[[1L]]))
     expect_equal(colnames(res[[1L]]), c("mz", "intensity"))
+
+    ## empty object LLLLL
+    ## long
+    tmp <- mm8_be_long[integer()]
+    expect_equal(peaksData(tmp), list())
+    ## blob
+    tmp <- mm8_be_blob[integer()]
+    expect_equal(peaksData(tmp), list())
+    ## blob2
+    tmp <- mm8_be_blob2[integer()]
+    expect_equal(peaksData(tmp), list())
 })
 
 test_that("peaksVariables,MsBackendSql works", {
-    expect_equal(peaksVariables(mm8_be), c("mz", "intensity"))
+    expect_equal(peaksVariables(mm8_be_long), c("mz", "intensity"))
 })
 
 test_that("intensity<-,MsBackendSql works", {
-    expect_error(intensity(mm8_be) <- 1:5, "replace")
+    expect_error(intensity(mm8_be_long) <- 1:5, "replace")
 })
 
 test_that("mz<-,MsBackendSql works", {
-    expect_error(mz(mm8_be) <- 1:5, "replace")
+    expect_error(mz(mm8_be_long) <- 1:5, "replace")
 })
 
 test_that("spectraData,MsBackendSql works", {
@@ -155,88 +166,88 @@ test_that("spectraData,MsBackendSql works", {
     expect_s4_class(res, "DataFrame")
     expect_equal(colnames(res), c("rtime", "mz"))
 
-    res <- spectraData(mm8_be, c("msLevel", "rtime", "mz"))
+    res <- spectraData(mm8_be_long, c("msLevel", "rtime", "mz"))
     expect_equal(res, spectraData(mm8_sps, c("msLevel", "rtime", "mz")))
 
     ## Arbitrary ordering
     idx <- c(4L, 12L, 100L, 14L)
-    be <- mm8_be[idx]
+    be <- mm8_be_long[idx]
     expect_equal(spectraData(be, c("msLevel", "rtime", "mz")),
                  spectraData(mm8_sps@backend[idx], c("msLevel", "rtime", "mz")))
 
     idx <- c(2L, 5L, 1L, 2L, 5L)
-    be <- mm8_be[idx]
+    be <- mm8_be_long[idx]
     expect_equal(spectraData(be, c("msLevel", "rtime", "mz")),
                  spectraData(mm8_sps@backend[idx], c("msLevel", "rtime", "mz")))
 })
 
 test_that("$<-,MsBackendSql works", {
-    be <- mm8_be
-    expect_error(mm8_be$spectrum_id_ <- "a", "not be")
+    be <- mm8_be_long
+    expect_error(mm8_be_long$spectrum_id_ <- "a", "not be")
     be$new_var <- "A"
     expect_true(any(spectraVariables(be) == "new_var"))
     expect_true(all(be$new_var == "A"))
 })
 
 test_that("reset,MsBackendSql", {
-    be <- mm8_be[c(5, 2, 10)]
+    be <- mm8_be_long[c(5, 2, 10)]
     be$add_var <- "B"
 
     be_res <- reset(be)
-    expect_identical(length(be_res), length(mm8_be))
+    expect_identical(length(be_res), length(mm8_be_long))
 })
 
 test_that("spectraNames,spectraNames<-,MsBackendSql", {
-    res <- spectraNames(mm8_be)
+    res <- spectraNames(mm8_be_long)
     expect_true(is.character(res))
-    expect_identical(res, as.character(seq_along(mm8_be)))
+    expect_identical(res, as.character(seq_along(mm8_be_long)))
 
-    expect_error(spectraNames(mm8_be) <- rev(seq_along(mm8_be)),
+    expect_error(spectraNames(mm8_be_long) <- rev(seq_along(mm8_be_long)),
                  "not supported")
 })
 
 test_that("filterMsLevel,MsBackendSql works", {
-    res <- filterMsLevel(mm8_be)
-    expect_equal(res, mm8_be)
+    res <- filterMsLevel(mm8_be_long)
+    expect_equal(res, mm8_be_long)
 
-    res <- filterMsLevel(mm8_be, msLevel = integer())
-    expect_equal(res, mm8_be)
+    res <- filterMsLevel(mm8_be_long, msLevel = integer())
+    expect_equal(res, mm8_be_long)
 
-    res <- filterMsLevel(mm8_be, msLevel = 1:2)
-    expect_equal(res, mm8_be)
+    res <- filterMsLevel(mm8_be_long, msLevel = 1:2)
+    expect_equal(res, mm8_be_long)
 
-    res <- filterMsLevel(mm8_be, msLevel = 3)
+    res <- filterMsLevel(mm8_be_long, msLevel = 3)
     expect_true(length(res) == 0)
 
-    tmp <- mm8_be
+    tmp <- mm8_be_long
     tmp$msLevel <- rep(1:2, 99)
     res <- filterMsLevel(tmp, msLevel = 1L)
     expect_true(length(res) == (length(tmp) / 2))
 })
 
 test_that("filterRt,MsBackendSql works", {
-    res <- filterRt(mm8_be)
-    expect_equal(res, mm8_be)
+    res <- filterRt(mm8_be_long)
+    expect_equal(res, mm8_be_long)
 
-    res <- filterRt(mm8_be, c(-Inf, Inf))
-    expect_equal(res, mm8_be)
+    res <- filterRt(mm8_be_long, c(-Inf, Inf))
+    expect_equal(res, mm8_be_long)
 
-    res <- filterRt(mm8_be, c(-Inf, 50000))
-    expect_true(length(res) == length(mm8_be))
+    res <- filterRt(mm8_be_long, c(-Inf, 50000))
+    expect_true(length(res) == length(mm8_be_long))
 
-    res <- filterRt(mm8_be, c(0, Inf))
-    expect_true(length(res) == length(mm8_be))
+    res <- filterRt(mm8_be_long, c(0, Inf))
+    expect_true(length(res) == length(mm8_be_long))
 
-    res <- filterRt(mm8_be, rt = c(1000, 2000))
+    res <- filterRt(mm8_be_long, rt = c(1000, 2000))
     expect_true(length(res) == 0)
 
-    res <- filterRt(mm8_be, rt = c(10, 20))
+    res <- filterRt(mm8_be_long, rt = c(10, 20))
     expect_true(all(res$rtime > 10 & res$rtime < 20))
 
-    res <- filterRt(mm8_be, rt = c(10, 20), msLevel. = 2)
-    expect_equal(res, mm8_be)
+    res <- filterRt(mm8_be_long, rt = c(10, 20), msLevel. = 2)
+    expect_equal(res, mm8_be_long)
 
-    tmp <- mm8_be
+    tmp <- mm8_be_long
     tmp$msLevel <- sample(1:3, length(tmp), replace = TRUE)
     res <- filterRt(tmp, rt = c(10, 20), msLevel. = 3)
     res_3 <- filterMsLevel(res, 3)
@@ -332,14 +343,14 @@ test_that("uniqueMsLevels,MsBackendSql works", {
 })
 
 test_that("backendMerge,MsBackendSql works", {
-    empty <- mm8_be[integer()]
+    empty <- mm8_be_long[integer()]
     res <- backendMerge(empty)
     expect_equal(res, empty)
 
-    spl <- split(mm8_be[1:10], 1:10)
+    spl <- split(mm8_be_long[1:10], 1:10)
     spl[[5]] <- empty
 
-    mm8_sub <- mm8_be[c(1, 2,3, 4, 6, 7, 8, 9, 10)]
+    mm8_sub <- mm8_be_long[c(1, 2,3, 4, 6, 7, 8, 9, 10)]
     res <- backendMerge(spl)
     expect_s4_class(res, "MsBackendSql")
     expect_true(length(res) == 9L)
@@ -380,14 +391,15 @@ test_that("setBackend works with MsBackendSql", {
     expect_error(setBackend(mm8_sps, MsBackendSql()), "required")
     tmpcon <- dbConnect(SQLite(), tempfile())
     expect_error(res <- setBackend(mm8_sps, MsBackendSql()), "dbcon")
-    res <- setBackend(mm8_sps, MsBackendSql(), dbcon = tmpcon)
+    res <- setBackend(mm8_sps, MsBackendSql(), dbcon = tmpcon, blob = TRUE,
+                      peaksStorageMode = "blob2")
     expect_equal(dbListTables(tmpcon),
-                 c("msms_spectrum", "msms_spectrum_peak_blob"))
+                 c("msms_spectrum", "msms_spectrum_peak_blob2"))
     expect_equal(mz(res), mz(mm8_sps))
     expect_equal(rtime(res), rtime(mm8_sps))
     expect_s4_class(res@backend, "MsBackendSql")
-
     dbDisconnect(tmpcon)
+
     tmpcon <- dbConnect(SQLite(), tempfile())
     res <- setBackend(mm8_sps, MsBackendSql(),
                       dbcon = tmpcon, blob = FALSE)
@@ -424,4 +436,27 @@ test_that("setBackend,Spectra,MsBackendSql works", {
     expect_true(length(processingLog(res)) > length(processingLog(ref)))
 
     dbDisconnect(con_test)
+})
+
+test_that("mz,MsBackendSql works", {
+    res <- mz(mm8_be_long)
+    expect_equal(res, mz(mm8_sps@backend))
+
+    a <- mm8_be_long[c(5, 1, 45, 1, 4)]
+    b <- mm8_sps[c(5, 1, 45, 1, 4)]
+    expect_equal(mz(a), mz(b))
+
+    expect_equal(mz(MsBackendSql()), IRanges::NumericList(compress = FALSE))
+})
+
+test_that("intensity,MsBackendSql works", {
+    res <- intensity(mm8_be_long)
+    expect_equal(res, intensity(mm8_sps@backend))
+
+    a <- mm8_be_long[c(5, 1, 45, 1, 4)]
+    b <- mm8_sps[c(5, 1, 45, 1, 4)]
+    expect_equal(intensity(a), intensity(b))
+
+    expect_equal(intensity(MsBackendSql()),
+                 IRanges::NumericList(compress = FALSE))
 })